Define helpers for working with environment
Right now these take the full env var name, but it might make more sense
to have them take the prefix/suffix separately. Unclear how much
complexity we want to put in the generation, and how much complexity we
want to put in the package code the generated code calls.