Add example generated code
This should guide the generator implementation. I think the pattern here
is pretty simple, but we'll see!

There are several TODOs and a bad note: we're sort of violating the
flag defaults pattern, since we default in this order:

```
CLI flag -> env variable -> zero value
```

So if the env variable is set, it will affect the help text.