Check in basic generation
This has rough edges. I don't like the treatment of the prefix, for
example: `LUKASVars` is awkward, unidiomatic go. We can afford to be
opinionated at first: this should be a single string with underscores as
separators.

+ `ONE_TWO` becomes `OneTwoXxx` in Go and `ONE_TWO_XXX` in the
  environment.
+ `one_two` behaves exactly the same way.

Also, pkg/env is currently a dependency of the generated code. Instead,
we should embed those functions: when you build `diego`, pkg/env should
get turned into a string, which should be appended to the generated
parser files. Using diego in a project shouldn't require importing it at
runtime.

Also, I think the root ./schema.json is superfluous: it's strictly
redundant with ./internal/scheam/schema.go.

Finally, but most importantly, the command line flag parsing is broken!