feat: generate parser from Go struct primary
This is an alternative to providing a JSON file as input: just give a
struct (with some kinda brittle tagging).

We probably want a dedicated field meta tag instead of using the `json`
name for each field.

I want to reorganize this. Whichever format you select, we should have
some "prepare" stage that yields

1. A `TemplateSchema`
2. A destination filename

And the overall pipeline should really be the same.